cloud: move code S3Config into separate handler
